requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database
SQLAlchemy
pydantic-settings
python-dotenv

# Authentication (JWTで今後使用)
python-multipart
email-validator
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1

# Testing
pytest
httpx